refactor: 날씨 서비스 인터페이스 도입 및 구현 분리
